requirements.txt
streamlit==0.79.0
streamlit-player-0.1.5
